move shouldInitDatabase test to db_test.go and t.Parallel()
