+++ lib/service/service_test.go
@@ -713,49 +713,3 @@ func getFreePort() (string, error) {
 
 	return l.Addr().(*net.TCPAddr).String(), nil
 }
-
-func TestTeleportProcess_shouldInitDatabases(t *testing.T) {
-	tests := []struct {
-		name   string
-		config DatabasesConfig
-		want   bool
-	}{
-		{
-			name: "disabled",
-			config: DatabasesConfig{
-				Enabled: false,
-			},
-			want: false,
-		},
-		{
-			name: "enabled but no config",
-			config: DatabasesConfig{
-				Enabled: true,
-			},
-			want: false,
-		},
-		{
-			name: "enabled with config",
-			config: DatabasesConfig{
-				Enabled: true,
-				Databases: []Database{
-					{
-						Name: "foo",
-					},
-				},
-			},
-			want: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := &TeleportProcess{
-				Config: &Config{
-					Databases: tt.config,
-				},
-			}
-			require.Equal(t, tt.want, p.shouldInitDatabases())
-		})
-	}
-}
